pyproject.toml
[tool.poetry]
name = "cheatsheet-generator"
version = "1.0.0"
description = "A tool to generate optimized cheat sheets from YAML hotkey definitions"
authors = ["Arkadii <arkadii@arkadhub.com>"]
readme = "README.md"
packages = [{include = "cheatsheet_generator", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11, <3.15"
pyyaml = "^6.0.2"
reportlab = "^4.4.3"
click = "^8.2.1"
pillow = "^11.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
black = "^25.1.0"
flake8 = "^7.3.0"
isort = "^6.0.1"
pre-commit = "^4.2.0"
pyinstaller = "^6.15.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src/cheatsheet_generator --cov-report=term-missing"
